project has been re-structured - divided onto pages and components
