+++ src/pages/Login/index.tsx
@@ -1,0 +1,7 @@
+export default function Login() {
+    return (
+      <div className="login">
+        there would be login page
+      </div>
+    );
+}
