+++ src/utils/isTrackKey.ts
@@ -1,0 +1,5 @@
+import {ALL_TRACK_KEYS} from "../data/plugs"
+
+export default function isTrackKey(value: string): value is TrackKey {
+    return ALL_TRACK_KEYS.includes(value as TrackKey);
+}
